Add back ability to use passwordless enable on EOS
This commit adds the ability to go into Privileged EXEC mode when no password is required by setting `enable: true`.

By default, in Arista EOS you aren't prompted for a password when entering Privileged EXEC (or privilege mode 15) mode. In PR #86, oxidized began always expecting a password.

An unquoted `true` in YAML is a boolean. So when `var(:enabled)` is `TrueClass`, we interpret this as meaning that the user wishes to go into Privileged EXEC mode and does not require a password.