Use the ceiling instead of the floor when calculating the desired number of jobs
If the number of jobs is large enough, using the floor won't guarantee completion
within the desired interval.
Example:
- interval: 3600
- average duration: 5
- number of nodes: 900
This gives 1.25 as the desired number of threads.
Flooring it means that nodes will only be collected once every ~75 minutes.

By using the ceiling and slightly over-estimating the desired interval is achieved