Sort the nodes by last update on reload
Reloading sets the node queue to the original order. If Oxidized was restarted,
collected a couple of nodes, and was then reloaded, the new queue would have a node
with a recent last end time.
Thus, no new nodes would be collected until the first one passed its interval.
Here the nodes are sorted on reload, to maintain the queue invariant.