Added EXPOSE and VOLUME directives
This describes the exposed ports and the volume according to the `README.md` documentation. 

Though it is possible to mount volumes and expose ports using `docker run`, for several docker management tools it is not possible to expose ports or mount volumes when they are not defined in the `Dockerfile`.