Issue identifying prompt with truncated hostname
The FortiOS device will truncate the hostname if it exceeds 24 characters. It will place a tilde (~) in the hostname, either at the end or before a hyphen. This causes a failure with prompt recognition and causes the backup to fail. This change adds a tilde as an identifying character in the prompt.