interpolate node variable values
Empty value was variably considered "" or nil, now it will always be ""
string.

Now "nil", "false", "true" strings are converted to their respective
types nil, false and true.

This also means we cannot have verbatim strings by that name, like if
your password was "false", you're shit out of luck. If this is the case,
we can add some __false__ hack or consider other similar solutions.

Fixes #500
Fixes #534

May break stuff with ""/nil changing now, but as it was not consistent
to begin with, I find it acceptable. Users now may need to manually
enter nil in some fields to regain old behaviour.