Hide db last update
Everyday the IPS URL db is updated and so the configuration is "changed". I believe this information spams the diff config and my team and I just decided to replace it by a unique string. Maybe instead of putting this in the :secret, we could create an other var or something that can be specified in the conf. Don't know what you think of this ;-)