Basic support for ArbOS from Arbor Networks.
ArbOS is the generic name for Arbor Networks (https://www.arbornetworks.com/) OS that runs on some of their devices.

Found and imported from https://github.com/claranet/oxidized . Credit to Claranet for the creation.

Locally validated on Peakflow SP v8.2+

The module is quite simple as the command 'config show' automatically disables pagination. There is a pagination-enabled version of the command as well.