fix: Allow Oxidized service to be stopped from services. (#1144)
There is an issue with oxidized and correctly terminatating when a user stops the service.

Unfortunately the root cause of this issue has yet to be identified, and it may
have to do with specific versions of Ruby gems.

As such, the workaround is to use a SIGKILL to kill the oxidized/puma service
instead of a SIGINIT.

Note: This does not solve the issue with manually starting the oxidized
daemon and using Ctrl + c to stop it. The solution here is to then
press Ctrl + Z, then manually "kill -9 <PID>".