Update boss.rb
Issues with some older boss switches that have the banner turned off.  They still display the banner but it is just scrolls and no pause.   The send "c" gets put at the prompt and you are now stuck.  An easy fix is to just send new line so you get a new prompt and the next command does not get corrupt.  Tested on about 100 BOSS switches and all working now.