Refactor manager
Main problem we're trying to solve is the silent failures. Requiring
files will never fail, one user had problem loading their souce http.rb,
which is obviously mandatory for them, but because it didn't fail, it
lead to a much harder to understand bug down the line, as we tried to
instansiate the 'nil' source.

Reason why I did it like this was laziness, because I didn't have to
care if or not the file exists, I always tried to load locally defined
models, and just return empty hash if I couldn't find any.

There may be other reasons and reliances that I've forgotten, but we'll
fix them as they come. Such is life without proper testing.