Refactor manager, take2
Make all the add_* return falsy valuem and make all the callers raise,
this has benefit that the caller has more information, like it can tell
what node we were trying to load when model failed to load.

We were only failing to do this in two places in the code, source and
plugin.