add parsing of configs
this adds the possibility to add a parse method to models.
it is used when serializing a node, extracting additional information
about CPU cores, RAM, etc. which is then stored into
a new field ":info" containing a hash of these values.
a list of keys and their meaning has to be defined, currently there are
these keys:

- manufacturer: hardware manufacturer
- name: product name
- version: software version
- cores: cpu cores
- ram: amount of RAM in bytes
- hdd: amount of non volatile memory in bytes

NB: i'm not sure if accessing the configuration requires locking as
it is done in yield_node_output in nodes.rb.