Create a model for Linksys SRW switches
This creates a model for Linksys SRW switches by sneaking out form the
menu mode and starting the lcli tool to be able to dump the config.

Due to the fact that we're navigating and escaping from a curses mode
not using the regular username/password entry procedures,
but rather chaining a bunch of expect to do the login and start the
lcli terminal mode.

This resolves #875

Signed-off-by: Anton Lundin <glance@acc.umu.se>