Support Aruba IAP unencrypted keys
When the script is run against a wireless switch, the command 'encrypt disable' is issued before retrieving the configuration. In IAPs, this command is not implemented but you can run 'show running-config ' with the suffix 'no-encrypt' and the result is similar. Unfortunately if you add this suffix in a wireless switch you will get an error message. So before sending the show run.. command we check wether is a wireless switch or an IAP by executing a 'show switchinfo' and then deciding which is the right command to retrieve the config.