hook/exec: preserve some environment variables
When executing a hook, the exec plugin removes all environment
variables. However, some variables can be useful to have: many
software crash if HOME is unset (including oxidized when we want to
use `oxs`) and PATH is useful to have for software not trying to
populate a PATH before invoking `exec*()` (unlike a shell).

Moreover, in case we want to use `oxs` in the hook, we also copy
`OXIDIZED_*` variables.