Update CHANGELOG.md
This doesn't work with our Cisco CSR hosts.  The ios.rb parser does not handle the ":" symbol in the prompt of the CSR device when oxidized logs in over ssh.  This update has been working fine for months in the production environment at my current job with librenms integration.

Created by basically combining the needed parts of ios.rb and iosxr.rb.