Update ip of example.com in ssh-test
Close #3145
The A-record of example.com was changed from 93.184.216.34 to
93.184.215.14, so our test failed.
I updated the IP-Adress in the ssh-test.
I could have added an automatic resolution in the test, but I don't
think the IP of example.com will change too often, so I kept it simple.
A comment should help to find the problem quickly next time.
I'm not documenting this in CHANGELOG.md as it is only relevant for
developpers and not for our users.