Update expected prompt (#3332)
This is to fix issue #3330 so multiline output won't break the module.

Some cumulus commands results do not end with \n, so the prompt gets
polluted by the trailing line of the previous command.
This is the case of `cat /etc/cumulus/ports.conf`, so we an echo to it to get an LF at EOF..

---------

Co-authored-by: Robert Cheramy <r.cheramy@netze-bw.de>
Co-authored-by: Michał Margula <mmargula@squiz.pl>