New Model for Unifi APs, switches, and AirOS APs
Here is a new model for several Ubiquiti devices. Probably if it has /tmp/system.cfg, and /etc/board_info it
has a decent chance of working. The hardest part was duplicating the output of the "info" command that is
seemingly an internal command in the shell, or just not available in the AIrOS versions I tried.

    Clif