New Model for Unifi APs, switches, and AirOS APs (#3431)
* New Model for Unifi APs, switches, and AirOS APs

Here is a new model for several Ubiquiti devices. Probably if it has /tmp/system.cfg, and /etc/board_info it
has a decent chance of working. The hardest part was duplicating the output of the "info" command that is
seemingly an internal command in the shell, or just not available in the AIrOS versions I tried.

    Clif

* Remove debug statement, opps. ;-)

* Update Supported-OS-Types.md

Added model for Unifi / AirOS APs and some Unifi switches

* Update CHANGELOG.md

---------

Co-authored-by: Robert Cheramy <r.cheramy@netze-bw.de>