input/ssh: hide Net::SSH errors
Revert to the log level of 0.33.0 for Net::SSH (:fatal, which is the default level of Net::SSH).
I could have disabled the use of ssh-agent with use_agent = false, but I'm not sure if someone uses an ssh-agent with oxidized. If we want to see the errors of Net::SSH when input.debug = false, we'll have to test if an agent is configured before opening a connection.
Fixes: #3574